+++ packages/ice-urql/src/runtime.tsx
@@ -1,8 +1,18 @@
 import type { RuntimePlugin } from '@ice/runtime/types';
 import React from 'react';
-import { Provider } from 'urql';
 import { RequestConfig } from "./types";
-import { createUrqlInstance, urqlInstances } from "./request.js";
+import { createUrqlInstance, getDefaultClient } from "./request.js";
+import {
+  AnyVariables,
+  DocumentInput,
+  Provider,
+  UseQueryArgs,
+  UseSubscriptionArgs,
+  SubscriptionHandler,
+  useQuery as useUrqlQuery,
+  useMutation as useUrqlMutation,
+  useSubscription as useUrqlSubscription
+} from 'urql';
 
 const EXPORT_NAME = 'urqlConfig';
 
@@ -19,9 +29,47 @@ const runtime: RuntimePlugin = async ({ appContext, addProvider }) => {
     createUrqlInstance(requestConfig);
   }
 
-  // addProvider(({ children }) => (
-  //   <Provider value={urqlInstances.default.client}>{children}</Provider>
-  // ))
+  addProvider(({ children }) => (
+    <Provider value={getDefaultClient()}>{children}</Provider>
+  ))
 };
+
+
+/**
+ * hook query
+ * @param args
+ * @returns
+ */
+export function useQuery<Data = any, Variables extends AnyVariables = AnyVariables>(
+  args: UseQueryArgs<Variables, Data>
+) {
+  return useUrqlQuery(args);
+}
+
+
+
+/**
+ * hook mutation
+ * @param args
+ * @returns
+ */
+export function useMutation<Data = any, Variables extends AnyVariables = AnyVariables>(
+  args: DocumentInput<Data, Variables>
+) {
+  return useUrqlMutation(args);
+}
+
+/**
+ * hook subscription
+ * @param args
+ * @param handler
+ * @returns
+ */
+export function useSubscription<Data = any, Result = Data, Variables extends AnyVariables = AnyVariables>(
+  args: UseSubscriptionArgs<Variables, Data>,
+  handler?: SubscriptionHandler<Data, Result>
+) {
+  return useUrqlSubscription(args, handler);
+}
 
 export default runtime;
